Time: 478 ms (5.03%), Space: 106.9 MB (6.13%) - LeetHub

+++ 1370-count-number-of-nice-subarrays/1370-count-number-of-nice-subarrays.cpp
@@ -4,8 +4,8 @@ public:
         for(auto&x : nums) {
             x%=2;
         }
-        map<long long, int> mp;
-        long long sum = 0;
+        map<int, int> mp;
+        int sum = 0;
         int ans = 0;
         mp[0] = 1;
         for(auto&x : nums) {
